backend/requirements.txt
# Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
groq>=0.4.0
duckduckgo-search>=4.1.0
pydantic>=2.5.0
